requirements.txt
absl-py==0.10.0
Augmentor==0.2.8
backcall==0.2.0
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
docopt==0.6.2
dominate==2.6.0
easydict==1.9
fake-useragent==0.1.11
faster-rcnn==0.1
filelock==3.0.12
fitlog==0.7.1
Flask==1.1.2
future==0.18.2
gdown==3.12.2
gitdb==4.0.5
GitPython==3.1.8
google-api-core==1.24.0
google-api-python-client==1.12.8
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
googledrivedownloader==0.4
goto-statement==1.2
graphviz==0.14.2
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.9
imageio==2.8.0
imgaug==0.4.0
imgviz==0.12.3
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.2.0
labelme==4.2.10
Markdown==3.3.1
MarkupSafe==1.1.1
matlab==0.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.32
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyblur==0.2.3
pycparser==2.20
pycurl==7.43.0.6
Pygments==2.7.1
pyparsing==2.4.7
pyPyrTools==0.2.3
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-complex-tensor==0.0.134
pytorch-ssim==0.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.1.0
Shapely==1.7.1
six==1.14.0
sklearn==0.0
smmap==3.0.4
tensor-sensor==0.1b6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.0
termcolor==1.1.0
torch==1.4.0
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.5.0
tornado==6.0.4
tqdm==4.44.0
traitlets==4.3.3
uncurl==0.0.10
uritemplate==3.0.1
urllib3==1.25.9
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
xerox==0.4.1
zipp==3.3.0
